# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nasdaq option chain",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tickers": {
      "get": {
        "summary": "Search Tickers",
        "operationId": "search_tickers_api_tickers_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/covered-calls/{ticker}": {
      "get": {
        "summary": "Get Covered Calls",
        "operationId": "get_covered_calls_api_covered_calls__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8,
              "title": "Ticker"
            }
          },
          {
            "name": "moneyness",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "itm",
                    "otm",
                    "all"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneyness"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoveredCallPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cash-secured-puts/{ticker}": {
      "get": {
        "summary": "Get Cash Secured Puts",
        "operationId": "get_cash_secured_puts_api_cash_secured_puts__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8,
              "title": "Ticker"
            }
          },
          {
            "name": "moneyness",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "itm",
                    "otm",
                    "all"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneyness"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSecuredPutPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashSecuredPutExpiration": {
        "properties": {
          "expiration": {
            "type": "string",
            "title": "Expiration"
          },
          "dte": {
            "type": "integer",
            "title": "Dte"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/CashSecuredPutContract"
            },
            "type": "array",
            "title": "Contracts"
          }
        },
        "type": "object",
        "required": [
          "expiration",
          "dte",
          "contracts"
        ],
        "title": "CashSecuredPutExpiration"
      },
      "CashSecuredPutPage": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "options_available": {
            "type": "boolean",
            "title": "Options Available"
          },
          "moneyness": {
            "type": "string",
            "enum": [
              "itm",
              "otm",
              "all"
            ],
            "title": "Moneyness"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "current_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Cents"
          },
          "current_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "stock_bid",
                  "chain_last_trade"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Source"
          },
          "stock_bid_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Bid Cents"
          },
          "stock_ask_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Ask Cents"
          },
          "market_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Session"
          },
          "is_real_time": {
            "type": "boolean",
            "title": "Is Real Time"
          },
          "quote_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Timestamp"
          },
          "last_trade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Trade"
          },
          "last_trade_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Trade Timestamp"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "chain_from_cache": {
            "type": "boolean",
            "title": "Chain From Cache"
          },
          "info_from_cache": {
            "type": "boolean",
            "title": "Info From Cache"
          },
          "history_from_cache": {
            "type": "boolean",
            "title": "History From Cache"
          },
          "risk_free_rate_pct_tenths": {
            "type": "integer",
            "title": "Risk Free Rate Pct Tenths"
          },
          "lows": {
            "$ref": "#/components/schemas/PeriodLows"
          },
          "expirations": {
            "items": {
              "$ref": "#/components/schemas/CashSecuredPutExpiration"
            },
            "type": "array",
            "title": "Expirations"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "options_available",
          "moneyness",
          "fetched_at",
          "current_cents",
          "current_source",
          "stock_bid_cents",
          "stock_ask_cents",
          "market_session",
          "is_real_time",
          "quote_timestamp",
          "last_trade",
          "last_trade_timestamp",
          "truncated",
          "chain_from_cache",
          "info_from_cache",
          "history_from_cache",
          "risk_free_rate_pct_tenths",
          "lows",
          "expirations"
        ],
        "title": "CashSecuredPutPage"
      },
...
      "CoveredCallExpiration": {
        "properties": {
          "expiration": {
            "type": "string",
            "title": "Expiration"
          },
          "dte": {
            "type": "integer",
            "title": "Dte"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/CoveredCallContract"
            },
            "type": "array",
            "title": "Contracts"
          }
        },
        "type": "object",
        "required": [
          "expiration",
          "dte",
          "contracts"
        ],
        "title": "CoveredCallExpiration"
      },
      "CoveredCallPage": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "options_available": {
            "type": "boolean",
            "title": "Options Available"
          },
          "moneyness": {
            "type": "string",
            "enum": [
              "itm",
              "otm",
              "all"
            ],
            "title": "Moneyness"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "current_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Cents"
          },
          "current_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "stock_bid",
                  "chain_last_trade"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Source"
          },
          "stock_bid_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Bid Cents"
          },
          "stock_ask_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Ask Cents"
          },
          "market_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Session"
          },
          "is_real_time": {
            "type": "boolean",
            "title": "Is Real Time"
          },
          "quote_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Timestamp"
          },
          "last_trade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Trade"
          },
          "last_trade_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Trade Timestamp"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "chain_from_cache": {
            "type": "boolean",
            "title": "Chain From Cache"
          },
          "info_from_cache": {
            "type": "boolean",
            "title": "Info From Cache"
          },
          "history_from_cache": {
            "type": "boolean",
            "title": "History From Cache"
          },
          "risk_free_rate_pct_tenths": {
            "type": "integer",
            "title": "Risk Free Rate Pct Tenths"
          },
          "lows": {
            "$ref": "#/components/schemas/PeriodLows"
          },
          "expirations": {
            "items": {
              "$ref": "#/components/schemas/CoveredCallExpiration"
            },
            "type": "array",
            "title": "Expirations"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "options_available",
          "moneyness",
          "fetched_at",
          "current_cents",
          "current_source",
          "stock_bid_cents",
          "stock_ask_cents",
          "market_session",
          "is_real_time",
          "quote_timestamp",
          "last_trade",
          "last_trade_timestamp",
          "truncated",
          "chain_from_cache",
          "info_from_cache",
          "history_from_cache",
          "risk_free_rate_pct_tenths",
          "lows",
          "expirations"
        ],
        "title": "CoveredCallPage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeriodLows": {
        "properties": {
          "d7_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "D7 Cents"
          },
          "d30_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "D30 Cents"
          },
          "d90_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "D90 Cents"
          },
          "d365_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "D365 Cents"
          }
        },
        "type": "object",
        "required": [
          "d7_cents",
          "d30_cents",
          "d90_cents",
          "d365_cents"
        ],
        "title": "PeriodLows"
      },
...
      "TickerListing": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name"
        ],
        "title": "TickerListing"
      },
      "TickerSearchResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/TickerListing"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "total",
          "results"
        ],
        "title": "TickerSearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}